+++ CHANGELOG.md
@@ -4,10 +4,13 @@
 
 New Features:
   - Get-RscWorkload cmdlet - Retrieves information about any supported workload type, with various filtering capabilities.
+  - Get-RscVmwareVm - Added -Relic switch. Use -Relic:$false to filter out Relics. -Relic will return only Relics. No usage of -Relic will return all (default operation). -Name parameter is now position 0, so you don't have to specify -Name.
 
 Fixes:
 
 Breaking Changes:
+
+ - Removed legacy Get-RscVsphereVm CS cmdlet. Please use Get-RscVmwareVm.
 
 ## Version 1.4